# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerBridge - Student Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  mentor: "mentor",
  mentee: "mentee",
  both: "both",
} as const;

export interface RegisterBody {
  /** Must be a .edu email address */
  email: string;
  name: string;
  password: string;
  role: RegisterBodyRole;
  districtId: number;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  mentor: "mentor",
  mentee: "mentee",
  both: "both",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  districtId: number;
  /** @nullable */
  districtName?: string | null;
  /** @nullable */
  bio?: string | null;
  subjects: string[];
  isVerified: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  mentor: "mentor",
  mentee: "mentee",
  both: "both",
} as const;

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  bio?: string | null;
  role?: UpdateUserBodyRole;
  subjects?: string[];
}

export type DistrictType = (typeof DistrictType)[keyof typeof DistrictType];

export const DistrictType = {
  high_school: "high_school",
  unified: "unified",
  elementary: "elementary",
  other: "other",
} as const;

export interface District {
  id: number;
  name: string;
  county: string;
  type: DistrictType;
  memberCount: number;
  openRequestCount: number;
}

export interface Tag {
  id: number;
  name: string;
  color: string;
  requestCount: number;
}

export type MentorshipRequestAuthorRole =
  (typeof MentorshipRequestAuthorRole)[keyof typeof MentorshipRequestAuthorRole];

export const MentorshipRequestAuthorRole = {
  mentor: "mentor",
  mentee: "mentee",
} as const;

export type MentorshipRequestStatus =
  (typeof MentorshipRequestStatus)[keyof typeof MentorshipRequestStatus];

export const MentorshipRequestStatus = {
  open: "open",
  matched: "matched",
  closed: "closed",
} as const;

export interface MentorshipRequest {
  id: number;
  authorId: number;
  authorName: string;
  authorRole: MentorshipRequestAuthorRole;
  districtId: number;
  districtName: string;
  title: string;
  description: string;
  tags: Tag[];
  status: MentorshipRequestStatus;
  /** @nullable */
  matchedUserId?: number | null;
  /** @nullable */
  matchedUserName?: string | null;
  createdAt: string;
}

export type CreateRequestBodyRole =
  (typeof CreateRequestBodyRole)[keyof typeof CreateRequestBodyRole];

export const CreateRequestBodyRole = {
  mentor: "mentor",
  mentee: "mentee",
} as const;

export interface CreateRequestBody {
  districtId: number;
  title: string;
  description: string;
  tagIds: number[];
  role: CreateRequestBodyRole;
}

export type UpdateRequestBodyStatus =
  (typeof UpdateRequestBodyStatus)[keyof typeof UpdateRequestBodyStatus];

export const UpdateRequestBodyStatus = {
  open: "open",
  matched: "matched",
  closed: "closed",
} as const;

export interface UpdateRequestBody {
  title?: string;
  description?: string;
  tagIds?: number[];
  status?: UpdateRequestBodyStatus;
}

export type CreateReportBodyReason =
  (typeof CreateReportBodyReason)[keyof typeof CreateReportBodyReason];

export const CreateReportBodyReason = {
  spam: "spam",
  harassment: "harassment",
  inappropriate: "inappropriate",
  fake_account: "fake_account",
  other: "other",
} as const;

export interface CreateReportBody {
  reportedUserId: number;
  reason: CreateReportBodyReason;
  /** @nullable */
  description?: string | null;
}

export interface BlockUserBody {
  blockedUserId: number;
}

export interface BlockedUser {
  id: number;
  blockedUserId: number;
  blockedUserName: string;
  createdAt: string;
}

export interface StatsOverview {
  totalUsers: number;
  totalMentors: number;
  totalMentees: number;
  totalDistricts: number;
  openRequests: number;
  successfulMatches: number;
  topTags: Tag[];
}

export interface DistrictStats {
  districtId: number;
  districtName: string;
  memberCount: number;
  mentorCount: number;
  menteeCount: number;
  openRequests: number;
  matchedRequests: number;
  topTags: Tag[];
}

export type ListDistrictsParams = {
  type?: ListDistrictsType;
  search?: string;
};

export type ListDistrictsType =
  (typeof ListDistrictsType)[keyof typeof ListDistrictsType];

export const ListDistrictsType = {
  high_school: "high_school",
  all: "all",
} as const;

export type ListRequestsParams = {
  districtId?: number;
  tagId?: number;
  role?: ListRequestsRole;
  status?: ListRequestsStatus;
};

export type ListRequestsRole =
  (typeof ListRequestsRole)[keyof typeof ListRequestsRole];

export const ListRequestsRole = {
  mentor: "mentor",
  mentee: "mentee",
} as const;

export type ListRequestsStatus =
  (typeof ListRequestsStatus)[keyof typeof ListRequestsStatus];

export const ListRequestsStatus = {
  open: "open",
  matched: "matched",
  closed: "closed",
} as const;