# lib/api-zod/src/generated/types/blockedUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerBridge - Student Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockedUser {
  id: number;
  blockedUserId: number;
  blockedUserName: string;
  createdAt: string;
}