# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerBridge - Student Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRequestBodyRole } from "./createRequestBodyRole";

export interface CreateRequestBody {
  districtId: number;
  title: string;
  description: string;
  tagIds: number[];
  role: CreateRequestBodyRole;
}