# lib/api-zod/src/generated/types/districtStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerBridge - Student Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Tag } from "./tag";

export interface DistrictStats {
  districtId: number;
  districtName: string;
  memberCount: number;
  mentorCount: number;
  menteeCount: number;
  openRequests: number;
  matchedRequests: number;
  topTags: Tag[];
}