# lib/api-zod/src/generated/types/mentorshipRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerBridge - Student Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MentorshipRequestAuthorRole } from "./mentorshipRequestAuthorRole";
import type { MentorshipRequestStatus } from "./mentorshipRequestStatus";
import type { Tag } from "./tag";

export interface MentorshipRequest {
  id: number;
  authorId: number;
  authorName: string;
  authorRole: MentorshipRequestAuthorRole;
  districtId: number;
  districtName: string;
  title: string;
  description: string;
  tags: Tag[];
  status: MentorshipRequestStatus;
  /** @nullable */
  matchedUserId?: number | null;
  /** @nullable */
  matchedUserName?: string | null;
  createdAt: string;
}