# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerBridge - Student Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterBodyRole } from "./registerBodyRole";

export interface RegisterBody {
  /** Must be a .edu email address */
  email: string;
  name: string;
  password: string;
  role: RegisterBodyRole;
  districtId: number;
}